requirements.txt
streamlit
openai
chromadb
sentence-transformers
python-dotenv
pandas
